Add `cargo zaplib serve`
Instead of an old Python 2 script. Closes #16.

I needed to serve files for the cross-browser CI testing, and I didn’t
want to fiddle with Python 2, so I figured I’d just do this now.

Will publish this crate when merged.
