+++ zaplib/cargo-zaplib/Cargo.toml
@@ -1,11 +1,17 @@
 [package]
 name = "cargo-zaplib"
-version = "0.0.1"
+version = "0.0.2"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/Zaplib/zaplib"
 repository = "https://github.com/Zaplib/zaplib"
 description = "Cargo extensions for Zaplib"
 
-[dependencies]
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 clap = { version = "3.0", features = ["derive"] }
+actix-web = { version = "4.0.0-rc.3", features=["openssl"] }
+log = "0.4.14"
+env_logger = "0.9.0"
+actix-files = "0.6.0-beta.16"
+rcgen = "0.9.1"
+openssl = "0.10.38"
