+++ zaplib/docs/src/contributing.md
@@ -6,7 +6,7 @@ There are 2 types of tests available. The browser test suite is more extensive, 
 
 * Running browser tests
   * Build the test suite: `cargo zaplib build -p test_suite`
-  * Have local server running: `zaplib/scripts/server.py`
+  * Have local server running: `zaplib serve`
   * Navigate to `http://localhost:3000/zaplib/web/test_suite/` and click `Run All Tests`
   * Test CEF by running `cargo run -p test_suite` and clicking `Run All Tests` (macOS Intel only, and first install CEF using `cargo zaplib install-deps --devel`).
 * Running jest tests
@@ -15,8 +15,5 @@ There are 2 types of tests available. The browser test suite is more extensive, 
 ## Updating the documentation
 
 * Build the website and watch for changes: `zaplib/scripts/watch_website_dev.sh` 
-* Run local server on port 4848: `website_dev/server.sh` 
+* Run local server: `zaplib serve website_dev/ --port 4848` 
 * Documentation sources are located at `zaplib/docs/src`
-
-
-
