--- zaplib/website_root/server.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -euo pipefail
-
-# Per https://stackoverflow.com/a/16349776
-cd "${0%/*}"
-
-PORT=4848 ../zaplib/scripts/server.py
